Update main README with release process overview
The main README for this repo is from very old release process
and has been entirely replaced by the thorough and complete steps
we keep up to date in the GitHub issue templates for this repo.

Replacing the content in this README with a high level overview and
then linking to the detailed templates for the complete process.

Signed-off-by: Jared Watts <jbw976@gmail.com>